feat: Add Welcome Page
This commit introduces a new Welcome Page to the application, which is displayed to unauthenticated users.

The key changes include:
- A new `WelcomePage` that serves as the entry point for unauthenticated users.
- The authentication form has been extracted from `AuthDialog` into a reusable `AuthForm` widget.
- The routing logic in `main.dart` has been updated to conditionally display either the `WelcomePage` or the main application based on the user's authentication status.
- A comment has been added to indicate where to insert code for operations to be performed after login but before the map is displayed.
