+++ lib/pages/welcome_page.dart
@@ -1,0 +1,60 @@
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import 'package:trainlog_app/l10n/app_localizations.dart';
+import 'package:trainlog_app/providers/auth_provider.dart';
+import 'package:trainlog_app/providers/settings_provider.dart';
+import 'package:trainlog_app/widgets/auth_form.dart';
+
+class WelcomePage extends StatelessWidget {
+  const WelcomePage({super.key});
+
+  Future<void> _onLogin(BuildContext context, AuthResult result) async {
+    final auth = Provider.of<AuthProvider>(context, listen: false);
+    final settings = Provider.of<SettingsProvider>(context, listen: false);
+    final success = await auth.login(
+      username: result.username,
+      password: result.password,
+      settings: settings,
+    );
+    if (success) {
+      // User is now authenticated, the main app will be shown.
+      // A comment can be added here for operations to be done after connection.
+    } else {
+      // Optionally, show an error message
+      ScaffoldMessenger.of(context)
+        ..hideCurrentSnackBar()
+        ..showSnackBar(
+          SnackBar(content: Text(auth.error ?? 'Login failed')),
+        );
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Trainlog.me App'),
+        centerTitle: true,
+      ),
+      body: Center(
+        child: SingleChildScrollView(
+          padding: const EdgeInsets.all(24),
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              Text(
+                AppLocalizations.of(context)!.loginToYourAccount,
+                style: Theme.of(context).textTheme.headlineSmall,
+              ),
+              const SizedBox(height: 24),
+              AuthForm(
+                type: AuthFormType.login,
+                onSubmitted: (result) => _onLogin(context, result),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
